move command client to virt-handler
