+++ pkg/virt-handler/cmd-client/generated_mock_client.go
